backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
async-timeout==4.0.2
asyncio==3.4.3
asyncpg==0.27.0
attrs==23.1.0
autopep8==2.0.2
bcrypt==4.0.1
bitarray==2.7.5
blinker==1.6.2
cacheout==0.14.1
captcha==0.3
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-log==0.4.0
cloudinary==1.33.0
cloudinary-cli==1.7.1
cytoolz==0.12.1
dataclasses-json==0.5.7
defusedxml==0.7.1
docstring-parser==0.15
eth-abi==4.1.0
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.4.0
eth-utils==2.1.1
Flask==2.3.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.4
Flask-Pydantic==0.11.0
Flask-SQLAlchemy==3.0.3
flask-talisman==1.0.0
frozenlist==1.3.3
gino==1.0.1
hexbytes==0.3.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
jsonschema==4.17.3
langchain==0.0.123
llama-index==0.4.37
lru-dict==1.2.0
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
multipledispatch==0.6.0
mypy-extensions==1.0.0
numpy==1.24.3
openai==0.27.2
packaging==23.0
pandas==2.0.2
parsimonious==0.9.0
Pillow==9.0.1
prettytable==2.4.0
protobuf==4.23.2
psutil==5.9.5
psycopg2-binary==2.9.6
pyaml==23.5.9
pycodestyle==2.10.0
pycryptodome==3.18.0
pydantic==1.10.9
Pygments==2.15.1
PyJWT==2.8.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
redis==4.5.5
regex==2023.6.3
requests==2.31.0
rlp==3.0.0
shortuuid==1.0.11
six==1.16.0
SQLAlchemy==2.0.16
talisman==0.1.0
tenacity==8.2.2
tiktoken==0.3.2
tk==0.1.0
tomli==2.0.1
toolz==0.12.0
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.6.3
tzdata==2023.3
ujson==5.7.0
urllib3==1.26.16
uvloop==0.17.0
watchdog==3.0.0
wcwidth==0.2.5
web3==6.4.0
websockets==11.0.3
Werkzeug==2.3.6
yarl==1.9.2
